+++ src/open_refinery/__init__.py
@@ -187,7 +187,7 @@ from .users import (
     valid_role,
 )
 
-__version__ = "1.17.0"
+__version__ = "1.18.0"
 
 __all__ = [
     "Factory",